+++ library/src/schemas/nullable/nullable.test.ts
@@ -10,9 +10,23 @@ describe('nullable', () => {
     const output = parse(schema, input);
     expect(output).toBe(input);
     expect(parse(schema, null)).toBeNull();
+
     expect(() => parse(schema, 123)).toThrowError();
     expect(() => parse(schema, false)).toThrowError();
     expect(() => parse(schema, undefined)).toThrowError();
     expect(() => parse(schema, {})).toThrowError();
+  });
+
+  test('should use default if required', () => {
+    const default_ = 'default';
+    const input = 'test';
+
+    const schema1 = nullable(string(), default_);
+    expect(parse(schema1, input)).toBe(input);
+    expect(parse(schema1, null)).toBe(default_);
+
+    const schema2 = nullable(string(), () => default_);
+    expect(parse(schema2, input)).toBe(input);
+    expect(parse(schema2, null)).toBe(default_);
   });
 });