+++ library/src/schemas/nullish/nullish.ts
@@ -4,23 +4,32 @@ import type { BaseSchema, Input, Output } from '../../types.ts';
  * Nullish schema type.
  */
 export type NullishSchema<
-  TWrappedSchema extends BaseSchema,
-  TOutput = Output<TWrappedSchema> | null | undefined
-> = BaseSchema<Input<TWrappedSchema> | null | undefined, TOutput> & {
+  TWrapped extends BaseSchema,
+  TDefault extends Input<TWrapped> | undefined = undefined,
+  TOutput = TDefault extends undefined
+    ? Output<TWrapped> | null | undefined
+    : Output<TWrapped>
+> = BaseSchema<Input<TWrapped> | null | undefined, TOutput> & {
   schema: 'nullish';
-  wrapped: TWrappedSchema;
+  wrapped: TWrapped;
+  get default(): TDefault;
 };
 
 /**
  * Creates a nullish schema.
  *
  * @param wrapped The wrapped schema.
+ * @param default_ The default value.
  *
  * @returns A nullish schema.
  */
-export function nullish<TWrappedSchema extends BaseSchema>(
-  wrapped: TWrappedSchema
-): NullishSchema<TWrappedSchema> {
+export function nullish<
+  TWrapped extends BaseSchema,
+  TDefault extends Input<TWrapped> | undefined = undefined
+>(
+  wrapped: TWrapped,
+  default_?: TDefault | (() => TDefault)
+): NullishSchema<TWrapped, TDefault> {
   return {
     /**
      * The schema type.
@@ -31,6 +40,15 @@ export function nullish<TWrappedSchema extends BaseSchema>(
      * The wrapped schema.
      */
     wrapped,
+
+    /**
+     * The default value.
+     */
+    get default() {
+      return typeof default_ === 'function'
+        ? (default_ as () => TDefault)()
+        : (default_ as TDefault);
+    },
 
     /**
      * Whether it's async.
@@ -46,13 +64,22 @@ export function nullish<TWrappedSchema extends BaseSchema>(
      * @returns The parsed output.
      */
     _parse(input, info) {
-      // Allow `null` or `undefined` values to pass
-      if (input === null || input === undefined) {
-        return { output: input };
+      // Get default or input value
+      let default_: TDefault;
+      const value =
+        (input === null || input === undefined) &&
+        (default_ = this.default) &&
+        default_ !== undefined
+          ? default_
+          : input;
+
+      // Allow `null` or `undefined` value to pass
+      if (value === null || value === undefined) {
+        return { output: value };
       }
 
       // Return result of wrapped schema
-      return wrapped._parse(input, info);
+      return wrapped._parse(value, info);
     },
   };
 }