+++ library/src/schemas/optional/optional.test.ts
@@ -10,9 +10,23 @@ describe('optional', () => {
     const output = parse(schema, input);
     expect(output).toBe(input);
     expect(parse(schema, undefined)).toBeUndefined();
+
     expect(() => parse(schema, 123)).toThrowError();
     expect(() => parse(schema, false)).toThrowError();
     expect(() => parse(schema, null)).toThrowError();
     expect(() => parse(schema, {})).toThrowError();
+  });
+
+  test('should use default if required', () => {
+    const default_ = 'default';
+    const input = 'test';
+
+    const schema1 = optional(string(), default_);
+    expect(parse(schema1, input)).toBe(input);
+    expect(parse(schema1, undefined)).toBe(default_);
+
+    const schema2 = optional(string(), () => default_);
+    expect(parse(schema2, input)).toBe(input);
+    expect(parse(schema2, undefined)).toBe(default_);
   });
 });