+++ library/src/schemas/optional/optionalAsync.test.ts
@@ -10,9 +10,23 @@ describe('optionalAsync', () => {
     const output = await parseAsync(schema, input);
     expect(output).toBe(input);
     expect(await parseAsync(schema, undefined)).toBeUndefined();
+
     await expect(parseAsync(schema, 123)).rejects.toThrowError();
     await expect(parseAsync(schema, false)).rejects.toThrowError();
     await expect(parseAsync(schema, null)).rejects.toThrowError();
     await expect(parseAsync(schema, {})).rejects.toThrowError();
+  });
+
+  test('should use default if required', async () => {
+    const default_ = 'default';
+    const input = 'test';
+
+    const schema1 = optionalAsync(string(), default_);
+    expect(await parseAsync(schema1, input)).toBe(input);
+    expect(await parseAsync(schema1, undefined)).toBe(default_);
+
+    const schema2 = optionalAsync(string(), () => default_);
+    expect(await parseAsync(schema2, input)).toBe(input);
+    expect(await parseAsync(schema2, undefined)).toBe(default_);
   });
 });